requirements.txt
streamlit
pillow
duckdb
pandas
langchain-google-genai
python-dotenv
plotly
lxml
